LPS-197087 As we do elsewhere in the Objects portlet, we should default to the "ID" field if the titleObjectFieldId is orphaned.
Note: We use get instead of fetch here, since we aren't planning to
gracefully handle the possibilty that it doesn't exist (this is not an
expected case), and if for some reason this possibility does occur,
we'll throw a more meaningful error message than a NullPointerException.
We should not be using fetch in cases where we immediately assume that
the object won't be null after having called fetch. get is better for
those cases since it will throw a more meaningful error message if the
object couldn't be found.