LPS-197299 Convert FDS filter machinery to TypeScript
This is to be able to leverage the types in @liferay/js-api for correctness
and to avoid future breaking changes.

Note that I had to stop at Filter.tsx because otherwise I'd be forced to
migrate the whole project and that's not the purpose of this PR. This lead
to some `@ts-ignore` comments and some usages of `any` that we will be able
to remove as we keep on migrating the FDS stuff to TypeScript.