LPS-197444 Make the wrapped request behave as a real one
Since we are delegating in the usual portal request dispatching from
ComboServlet, we need to make the rest of the portal think that the
requests from ComboServlet are real.

Because we are using RequestDispatcher.include, the request that the portal
sees (unless we do something to remediate it) is for /combo, not for the
subresources the combo is trying to gather together.

In this commit we override some of the wrapped request methods to make
portal think the request is a normal one.