LPS-199370
Normally, the container response filter will handle cleaning up the transaction.

If a RuntimeException or Error is raised in an exception mapper, filters stop being processed, so we'll need to clean up the transaction in a phase interceptor.

If a RuntimeException is raised in a phase interceptor, phase interceptors stop being processed, so we'll need to clean up the transaction in a fault observer.

For now, we'll need to hope that nobody writes a phase interceptor that raises an Error, because these changes won't handle that case.