LPS-202718 Use index keys since current sub implementation does not support named keys
With the latest sub update we have (accidentally) remove support for named
keys in translations. It is very difficult to add this functionality since we
need to deal with mixed arrays + objects + react nodes, and it did not add
much value, so for now we can just rely on indices.