LPD-21814 gradle-plugins-workspace: fixes createDockerContainer test failure
Since the mustRunAfter and finalizedBy relationship now forces certain tasks to
run first (correctly), the dependent tasks of "removeImage"
("removeDockerContainer" and "stopDockerContainer")  already ran at the
beginning of the execution phase since it is declared that
"createDockerContainer" must run after "removeDockerContainer".

Since we can't force those tasks to re-run, the solution here is to create
dedicated stop and remove tasks that mirror the originals, and use these as
dependencies instead.