LPD-24699 Remove the plugin time leftover api PortalLifecycle
# breaking

## What portal-impl/src/com/liferay/portal/servlet/PortalSessionCreator.java

Removed PortalSessionCreator

## Why

The logic is moved into PortalSessionListener using DependencyManagerSyncUtil.registerSyncCallable()

----

# breaking

## What portal-impl/src/com/liferay/portal/servlet/PortalSessionDestroyer.java

Removed PortalSessionDestroyer

## Why

The logic is moved into PortalSessionListener using DependencyManagerSyncUtil.registerSyncCallable()

----

# breaking

## What portal-kernel/src/com/liferay/portal/kernel/deploy/hot/HotDeployEvent.java

Removed addPortalLifecycle(PortalLifecycle) and flushInits()

## Why

These are plugin time leftover apis which are not used anymore.

----

# breaking

## What portal-kernel/src/com/liferay/portal/kernel/deploy/hot/HotDeployUtil.java

Removed registerDependentPortalLifecycle(String, PortalLifecycle)

## Why

This is plugin time leftover api which is not used anymore.

----

# breaking

## What portal-kernel/src/com/liferay/portal/kernel/servlet/filters/invoker/InvokerFilter.java

Removed base class BasePortalLifecycle

## Why

This is plugin time leftover api which is not used anymore.

----

# breaking

## What portal-kernel/src/com/liferay/portal/kernel/util/BasePortalLifecycle.java

Removed class BasePortalLifecycle

## Why

This is plugin time leftover api which is not used anymore.

----

# breaking

## What portal-kernel/src/com/liferay/portal/kernel/util/PortalLifecycle.java

Removed class PortalLifecycle

## Why

This is plugin time leftover api which is not used anymore.

----

# breaking

## What portal-kernel/src/com/liferay/portal/kernel/util/PortalLifecycleUtil.java

Removed class PortalLifecycleUtil

## Why

This is plugin time leftover api which is not used anymore.

----

# breaking

## What util-taglib/src/com/liferay/taglib/servlet/JspFactoryServletContainerInitializer.java

Removed class JspFactoryServletContainerInitializer

## Why

This has been deprecated and not used.

----