LPD-23771 In *SerDes.java: Updated toJSON Serialization Method
# breaking

## What modules/util/portal-tools-rest-builder/src/main/resources/com/liferay/portal/tools/rest/builder/dependencies/client_serdes.ftl

This commit revises the toJSON method within *SerDes.java across all rest-client modules. Previously, when serializing an entity with an array field, the array was parsed as a string. With this update, the serialization behavior has been enhanced to properly handle arrays, now utilizing a new private static method _toJSON for improved serialization.

## Why

Array fields need to be correctly represented in JSON format.

----