LPD-30519 Articles may have empty/null content
It is possible to create a journal article with empty content. The
main use case is when defining default values for a structure; that is
implemented by creating a hidden journal article containing those
default values. In this case, it is perfectly sensible for the content
field to be empty/null.