LPD-33708 Rework CSP rules to gather them in one single SF check
Since CSP checks ares a bit specialized, we put them in a CSPComplianceCheck instead of
spreading them in different checks.

This is because we may need to add more things in the future to comply with CSP and
because it may be useful to have an isolated class to make all CSP related checks since
it can be used outside of the regular SF process in the future.