LPD-25508 Delay company creation via OSGI configs, until initial request
This company creation mechanism (via OSGI config), was intended to
be used with DBPartition enabled. But, with the current state of
DBPartition, creating a new partition requires the default company
to first be fully ready and populated, since it needs to copy over
certain tables from the default company to the new partition.

And so, delaying the company creation until the initial request
should help to ensure that the default company is fully ready and
populated, before attempting to create a new company and partition.