LPD-45266 Avoiding pollution of permission thread local
During the fragment rendering there are some scenarios in which the
request being passed to RESTClient is not authenticated. When that
happens the PermissionThreadLocal is set with the guest user inside the
REST call overriding the previous user.

The issue affects cases where REST Client is run several times within
the same thread.

With these changes we restore the original permission checker after the
REST Call to avoid side effects.