LPD-51750 Leave room in JS bootstrap process for Liferay global object
We need to populate Liferay global object in two phases: one before AUI is
loaded, the second one after that. That is because AUI relies on some parts of Liferay global object, but at
the same time the Liferay global object relies on AUI.

In this commit we leave room for another DynamicInclude that will declare
the Liferay global object, so that we can remove it from top_js.jspf.