LPD-55605 Remove invalid test
I have investigated this and have found that we never required the Source
Code URL to be a valid URL.

Doing this for all types would require a considerable change in the CET
framework and I can't see any benefit other than breaking customer
installations since the source code URL field is purely informative so it's
up to the customer to put a valid URL or maybe a local path or anything
that can help them to review their stuff.

In any case DXP won't use that URL for anything.