LPD-64397 Re-add curvesapi library for POI shared dependencies
In the past, the curvesapi library was added as a portal
dependency to fix this exact bug (LPS-125731). However, this
was eventually removed by LPS-144549.

The removal of the dependency resulted in re-introducing the bug
and this was most likely unintentional.
- LPS-144549 was part of a task to clean up portal libs
- There was no test that covered the bug's use-case.

----

The Apache POI library (poi-ooxml) has a dependency on the
curvesapi library and relies on this library when reading vsdx
files or files that contain vsdx diagrams. It can also be used to
extract textual context from the diagrams, which we utilize when
using Tika.

Reference: https://poi.apache.org/components/diagram/index.html

The vsdx file format is commonly used by MS Visio for creating
graphical diagrams. These diagrams may also be embedded in
documents such as powerpoint presentations (.pptx).