LPD-57326 Export a helper from node-scripts to hashify theme files
We need to hashify themes' CSS and JS files so that they can benefit from
infinite caching. Since we need to do it from several projects I've created
an `exports` folder in `node-scripts` so that we can require it from the
`gulpfile.js` of theme projects.

I tried to make it automatic by passing a template `gulpfile.js` from
`node-scripts` and deleting the ones in the project folders, but it didn't
work because all our Java tools use `gulpfile.js` files as markers and even
look inside them so we cannot delete or change them without affecting the
build.